+++ test/usecases/repo_usecases_test.dart
@@ -203,6 +203,10 @@ class MockRepoRepository implements RepoRepository {
   }
 
   @override
+  Future<Either<Failure, PullRequest>> createPullRequest(
+    String owner, String repo, Map<String, dynamic> body) => throw UnimplementedError();
+
+  @override
   Future<Either<Failure, void>> mergePullRequest(
     String owner,
     String repo,